+++ components/@dopt/react/modal/sandbox/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dopt/react-modal-sandbox
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [6cf837db3]
+  - @dopt/react-modal@2.0.0
+  - @dopt/react-theme@1.0.0
+
 ## 0.0.9
 
 ### Patch Changes